feat: write file if changes

+++ packages/frourio/src/writeRouteFile.ts
@@ -2,6 +2,8 @@ import fs from 'fs'
 import { Template } from './buildServerFile'
 
 export default ({ filePath, text }: Template) => {
+  if (fs.existsSync(filePath) && fs.readFileSync(filePath, 'utf8') === text) return
+
   fs.writeFileSync(filePath, text, 'utf8')
   console.log(`${filePath} was built successfully.`)
 }
